+++ R/utils_model_family.R
@@ -18,7 +18,7 @@ get_glm_family <- function(fit) {
   } else {
     # "lrm"-object from pkg "rms" have no family method
     # so we construct a logistic-regression-family-object
-    if (mc == "lrm")
+    if (mc %in% c("lrm", "polr"))
       faminfo <- stats::binomial(link = "logit")
     else
       # get family info